Moved dropout_input into utility module
